requirements_for_tests.txt
-r requirements.txt
coverage==3.7.1
coveralls==0.4.2
httmock==1.2.2
mock==1.0.1
nose==1.3.3
pep8==1.5.7
PyHamcrest==1.8.0
Flask-Testing==0.4.2
